Refactorizes the code for a better project structure
The source code has been refactorized, mainly, to group the different
components of the project (components, models and services) into
different subfolders of app.

In addition, the video and polyp mock arrays are no longer accesses
directly, but through Angular service classes. This change prepares
the application for the replacement of mock classes with real HTTP
requests.