Shows the polyps of an exploration
This commit changes the list of all polyps by the list of polyps of an
exploration since the polyps are now related to their exploration.
In addition, to adding a new polyp to an exploration, now that element is
incorporated into the exploration.
The service of listing all polyps is also removed since it's not necessary.