Adds relationship between polyp and video, and get the relationship
This commit adds the entity PolypRecording to relate the polyps with
the videos. Creates a service to obtain this relatioship. However, the
functionality is not complete since it's necessary to recover the polyp
and the video, but only the URL of the resource is known to obtain the
polyp or video. Therefore, in the video editing view it can see where
the polyp starts and ends, but the name of the polyp is not displayed.