Adds edit a video
In the view, the status of the button changed to edit or save the editing,
for this, the input elements are enabled or not. Also adds the functions
to execute the editing of a video. Also modifies the models to receive an
IdAndUri or send a string with the identification of the exploration.