Adds create a patient
This commit adds the functionality to create a patient. In the view is added
a button that opens a modal window to collect the necessary data. This modal
window is validated so that not enter null fields.
The model and the mapper are included for the Patient object.
In addition, as this entity also has an enumerared, utility functions are
created to be used by any entity that has been enumerated.

It is necessary to modify the exploration to include a patient.