Modifications in creating and editing exploration
Now the exploration is related to a patient. To do this, the attribute is added in
Exploration and ExplorationInfo.
When creating the exploration, the user has to indicate a patient. The view shows the
patients that are searched by the patientID while the user is writing in the text
entry (autocomplete). Check that the patient is correct before making the creation.
It is not allowed to modify the patient of the exploration, but this object is necessary
to recover it when the exploration is obtained. For this reason, a function is also
added to search a patient by the identification.