Adds assign a modifier to a video
Adds the form to indicate a modifier in a video in the modification tab in edit
video. This is done from a new component. Adds both the service and the models
to assign a modifier to a video.
The conversion of number to string is necessary in several components, therefore
it is extracted to a pipe.