Adds notification and error handler services
A notification service was added. This service can be used by other
services to notify successful, information, warning or error events.

Currently, when a notification is sent through the notification service
a toast message is displayed in the user interface. The toast service
used does not work well with error messages, and should be reviewed.

In addition, a custom error handler has been added. This error handler
captures PibaError and HttpResponseError errors and generates an error
notification in the notification service.