Changes in patient search
The coded patient identifier is not recognized by the user. However, this attribute
can not be decoded. Therefore, to show the readable data, the patient identifier is
shown as the user search.
In addition, more information about the patient is added in the view. A pipe is added
to calculate the age.