Adds create user for administrador role
Creates a component to manage users by an administrador. For now, adds a link
in the lateral menu to access this management. Also creates a service,
at the moment, to create users.
The Users entity refers to the data that will be stored in the database, while
the User entity as the user logged into the system.