Adds the edition of an user
In the user edition, the same modal window is used as in the creation, calling
the corresponding service. The only field that can be edited is the password, so
in the edit modal only this field is displayed.

In addition, the password confirmation is added in the creation and editing, ensuring
that the same password is introduced two times.