Adds attribute to know if a video contains text
Adds a new attribute to the Video entity to specify whether the video has text
or not. This attribute can be indicated when uploading a video, or can be
modified when editing a video.
Also fixes the editing of a video. When selecting the editing of a video, all the
inputs of all the videos were enabled. Now, only the inputs that you want to
edit are enabled.