Adds remove user and confirmation for this deletion
Adds the functionality to delete a user, which is searched by the login.

In addition, a new component is added to do the deletion confirmation.
The confirmation is a modal that can be activated by any component that
needs it. In this modal the deletion can be canceled or confirmed. It
can also receive a string with details of what is being deleted.