Adds progress bar when uploading a video
A modal with a progress bar is displayed while the video is being
uploaded and is hidden at the time the upload is completed. At that
moment the video begins to be processed. The data displayed next to the
progress bar in the modal is the upload percentage, the estimated time
remaining and the upload speed (in MB/s). To obtain this data, the
HttpEvent class is used, which unifies all possible events on the
response stream. If the type of this class is UploadProgress (while
uploading), the data is calculated. When the type changes to Response,
the video upload is completed.

In addition, a function (getMapVideoInfo) is created in the video
service that allows calling the mapVideoInfo function from outside the
service. This is necessary in the exploration component at the time when
HttpEventType changes to Response to obtain the video that has just been
uploaded.