Fixes canceling editions
When editing explorations, idspaces, polyps and users, the modal dialog
were writting directly to the objects in the list. So, if the user cancels
the editing operation, objects remain modified in the front-end. This commit
creates a clone of the editing object before bringing up the modal dialog,
fixing this undesired behaviour.