Refactors video upload
This commit refactors video upload functionality in order to
keep the exploration component independent of HttpEvent. This
is done by expanding the createVideo method in the videos service
in order to receive custom callbacks to monitor the uploading
progress. Such callbacks are defined in the exploration component
and are in charge of updating the user interface with the progress
status, but the HttpEvent is hidden for the components layer.