Fixes the search for exploration by patient
The IdSpace object was used in both the search field and the field to
add exploration. This caused that when adding a exploration, the IdSpace
to be searched for was also modified. This commit created two differentiated
objects, so the modification of one does not cause modifications in the other.