Changes the pagination of the explorations
The default pagination of Clarity is changed by the pagination made in
the backend. Now, only the data of the current page is obtained. Two
fields are added to the component: the page size (currently 15) and the
current page, which will be communicated to obtain the corresponding
explorations. The total number of explorations (both the patient's and
the total ones) is collected from the "X-Pagination-Total-Items" header.
This is necessary to calculate the number of pages that will be
displayed. The complete response of the request must be accessed with
the "{observe:'response'}" option to read the header. Because of this,
the service methods have been adapted, returning an ExplorationPage
object. ExplorationPage is a new class that contains the total number of
explorations and the explorations that will be displayed on the
indicated page. At all times it is verified which explorations are being
paginated (those of the patient or the totals) to make the corresponding
request.

To know when the datagrid changes (mainly when changing the page),
clrDgRefresh is used, which emits its current state.

In addition, the smart iterator *clrDgItems is replaced by a simple
*ngFor because *clrDgItems would paginate the subset of the data that
has already gone through that. For this reason, the
ExplorationComparator (which was necessary to sort the explorations by
date in the table) has been removed since this required the smart
iterator of Clarity.