Adds pagination to the images
Adds pagination in the list of images of the gallery.

When editing an image, there is also a hidden gallery to be able to
scroll through the rest of the images. This information is also
paginated, so it is controlled when changing pages and retreiving
new images.
In this view the information of the polyp recording is also shown,
this data is calculated with the frames per second
of the video.

The attribute frames per second is also added to calculate the
second of the polyp recording to which the images corresponds.