Disables add and edit button when polyp name is in use
When a new polyp is added, both from the video editing view and from
the polyp view, if the name of the polyp is being used the button is
disabled and an error message is displayed.

When editing the polyp the name of the polyp is saved. The button is
only disabled if the name matches that of another polyp and is not
the same as the one it already had (the saved name).