Adds the modifiers view
A new component has been created to manage the modifiers. The view of
modifiers (which can only be seen by administrators) is accessible
from all views. In this view, the name of the modifier and the
buttons of the add, edit and delete actions (not yet implemented) can
be seen.