Adds the "ENDOSCOPIST" role
This commit adds the endoscopist role which can perform the same actions
as the USER role. In addition, the endoscopist is the only one who can
create or remove a polyp location in the image and remove images. Within
each image, the endoscopist only sees these actions, without showing him
the polyp information.

Now, the galleries are accessible for all user roles but only the ADMIN
can edit them. The button to delete galleries is removed since the
removal of these is not supported.