Adds a warning message when polyp location has not been saved
This commit adds a modal that opens when it wants to change the image
(both to the right and to the left) and there is a polyp location that
has not been saved. This warning is only shown to the endoscopist, since
this role is the only one that can save locations. To know which
function to call (toRight or toLeft) the variable right is created. This
variable is set to true when the right arrow is pressed.