Adds a filter to see the images and optimizes assigning location to the polyps
Adds a filter to see the images in the gallery based on this filter.

Optimizes the application when the polyps are located. Instead of having
images preloaded according to the pagination, it is get a list of the
identifiers of the images according to the selected filter. When scrolling,
only the next or previous required imaged is requested.

Renames some function to facilitate the legibility of the code. It also
resolves unwanted behaviors with the modal warning that the location has
not been saved, por example, this modal never hidden by clicking continue.