Adds the polyp location progress in the gallery
In each gallery a progress bar is displayed indicating the located
images, the total images and the percentage completed. This bar will
only be shown if the number of images is greater than 0. To do this, a
new field is added to ImagePage, called locatedImages. This field
contains the total number of located images in the gallery.