Disables add and edit button when video title is in use
If the video title is already being used when a new video is added, the
add button is disabled and an error message is displayed.

When editing, the video title is saved. The button is only disabled if
the title matches that of another video and is not the same as the one
it already had (the saved title).