Adds autocomplete the observation to remove in the image deletion
To facilitate writing the observation to remove an image, the
autocomplete is added. When the number of characters entered is greater
than three, the previously inserted observations are consulted and those
that start with the characters entered are shown. The options shown can
be selected. To not use any of those options it is pressed "Enter".