Changes the pagination of explorations to search by specific page
Adds an input next to the pagination that allows to enter the desired
page. When the page is entered, it is checked if it is a valid page, if
not, an error message is displayed. If it is a valid page, a delay of
500 ms is added so that the user has time to write, avoiding making a
request for each character entered. In order to correctly detect
changes, pagination has been adapted using the clarity component
"ClrDatagridPagination".

In addition, as only changes are detected when the page changes, to
prevent the input from having a number that does not correspond to the
page selected (when performing a search from page 1), it is assigned the
value 1 directly from the component.