Changes the pagination in images gallery to search by specific page
Adds an input next to the pagination that allows to enter the desired
page, as in the pagination of explorations. For this modification the
ClrDatagridPagination component of clarity is used and it is checked if
the page that is entered in the input is correct. To prevent the input
from having a number that does not correspond to the selected page when
the filter changes, it is assigned the value 1 directly from the
component.

In addition, the manually inserted spinner (used while images are being
loaded) has been changed by the spinner provided by Clarity (with the
use of clrDgLoading) to facilitate the implementation of pagination.