Improves video progress management and capturing
This commit improves video progress management and capturing
in order to help user to better adjust the start and end times
of video annotations. In concrete:

1. The backward and forward buttons always place video current
time in the .000 and .999 millisecond, respectively. With this
modification, the user will see the first and last frame of
the second.

2. The start and end time capture buttons first pauses the video
and places its current time in the .000 and .999 millisecond,
so the user will see the first and last frame of the
captured second, respectively.

In addition this commit fixes the current second display in the
video component.