Adds the capture of images from a video
This commit adds the functionality of making snapshots of a video. To
make the snapshot, a button is added in the video (camera button) that
opens a modal that shows the image corresponding to the frame when the
button is pressed. If the video is being executed when the capture is
made, the video is paused. The snapshot can only be done by the role
ENDOSCOPIST.

Canvas is used to obtain the image data and display it. Since the video
used to write information to the canvas is in another component, then
the canvas becomes tainted. A tainted canvas can still be written to,
but it cannot be read from. Attempts to read data back from a tainted
canvas will result in a security error. To fix this error, the
cross-origin attribute with the anonymous value (crossorigin =
"anonymous") is used in the video.

In the modal, two fields are added to indicate the gallery in which the
image will be saved (obligatory field) and the polyp that appears in the
image (optional). A search was implemented in these two fields without
case sensitive. A checkbox (checked by default) is added next to the
polyps that allows showing the polyps that already exist in that frame
(checkbox checked) or all the polyps of that exploration (checkbox
unchecked).

In addition, the submit button that saves the image in a gallery is
disabled if the gallery or the polyp is incorrect or if the input
doesn't match the name of the previously selected gallery/polyp.