Updates the images of the galleries to use the polyp field of these
To show the information of the polyp, the polyp of the image is used
instead of the corresponding polyp recording. As now the images don't
have to contain a polyp, it is checked if the polyp is defined to show
the information. In addition, if the image does not contain a polyp,
it is not allowed to locate it.