Adds the download of an image from the video
This commit adds a button to the snapshot modal that allows downloading
the image obtained in that capture. To allow the download to all users,
the camera button of the video that allows making the snapshot is
enabled for everyone. Now it is checked in the modal if the role is
"endoscopist" to show the data of sending the snapshot to a gallery.

By default, the name of the image to be downloaded is the id of the video
followed by the frame number.