Adds show/hide the visualization of polyp locations in the gallery
This commit adds a checkbox in the gallery (next to the filter) that
allows to show the location of the polyps (checkbox checked) or hide it
(checkbox unchecked). This information is sent to the image (as a
queryParam) when it is clicked on, which allows to keep the
visualization of the polyp location (shown or hidden) at the moment of
displaying the image.

In addition, if the location is not showing in the image, the save and
clean buttons are disabled, as well as not allow to draw any location on
the image.