Adds three new fields to the table where the galleries are displayed
Now, in the galleries component, three new data are added for each
gallery: images with polyp, percentage of located polyps and total
images. These data are shown next to those shown before (name and
description). To obtain them, a request is made for each gallery that
obtains the ids of all the images it contains, the total number of
images, the number of images with located polyp and the number of images
with polyp. This data has been added to a map (with key id of the
gallery) to be obtained at any time without making another request.

In addition, as the number of requests is high, the datagrid where the
information is displayed will be in a loading state until all requests
are made.