Fixes the addition of a new polyp after creating another polyp
In the video editing view, creating a new polyp and then trying to
create another showed an error indicating that the field should not be
empty. This is because the error control of the previous creation form
was maintained. To solve this, the input where the polyp is inserted was
marked  as pristine again.