Adds a shortcut to save the location of the polyp
Now, to save the location of the polyp it is only necessary to press the
"s" key. The same checks are made as when pressing the save button. In
addition, a new check is added when saving using the shortcut or the
button, which does not allow saving if the location of the polyp is the
same as the stored one.