Fixes delay when drawing the location of a polyp
When inserting the download button of an image, a delay was introduced
when painting the location of a polyp. When clicked on the image and
scroll to draw, it takes time to see the box and latency is observed.

This commit returns the fluidity to the movements when drawing box on
the image. The reference to the image is only assigned when the user
clicks on the download button, avoiding so much delay to paint.