Adds the download of full image galleries
A button is added to the actions of the galleries that allows to
download them. Pressing the download button opens a modal window that
allows to select the download filter and whether or not to add
the location of the polyp. Two validations are done:
 1. The filter should always return images. In case the filter returns 0
 images, the download button is disabled.
 2. If the filter corresponds to non-located images, the checkbox that
 allows to select whether to draw the location is disabled and the value
 is marked as false.

The download is done through the HTML anchor element, in which the REST
URL is passed through the href attribute.