Fixes the galleries view when is empty or a new one is added
This commit fixes in the gallery view:

1. When there are no galleries, pagination is not shown to avoid errors.

2. When there were no galleries there was a spinner loading all the
time. This commit fixes this error and now a message is displayed
indicating that there are no galleries.

3. Before, when a new gallery was added, it was not automatically
displayed in the view. This commit fixes this error, which occurred when
trying to obtain the number of images of this new gallery and returned
an undefined result.