Adds colors to areas with polyps and modifications in the video
Now the regions of the video that contain polyps or modifiers are
reflected with a color in the video bar. The parts that match polyps and
modifiers are represented with another color, the result of mixing the
previous two. In order for the modifiers to be updated when one is
added/deleted, an output variable is created that emits the modifiers to
the main component. As the modifier component is not loaded until the
modifier tab is pressed, the first initial search for them is performed
independently of the output variable.

In addition, legends and checkboxes are added below the video. The
legends indicate the colors used along with what they represent.
Checkboxes allows to indicate if it is wanted to see polyps, modifiers
or both (by default, both are activated).