Fixes the video bar
This commit fixes the following errors in the video bar:

1. When pressing stop, the progress bar remained at the moment when stop
was pressed. Now the progress bar returns to the beginning.

2. When the video was playing and the selected speed was 1, when trying
to backward it remained in the same second. Now it is verified if the
video is playing and in that case a second is subtracted.

3. When the video was stopped, when using the forward and backward
buttons the progress bar did not move. Now these buttons work correctly.