Adds password recovery
This commit adds the password recovery system.
The visual part of that is made on login component, both creating the
request and creating the new password.
Connection to backend is defined on the users service.
Also in the app-routing module, there is a new route like
"/loginrecovery?uuid=" that user gets from the email sent when requests a
recoovery, and that redirects him to create his new password.
Creating a request can be made using login or email.
Creating a new password requires to introduce it two times and that both
are coincident.