Adds observation field to image
This commit allows to add observations to the images obtained by
capturing the video before being saved in a gallery. This action can
only be done by the endoscopist. These observations are shown within
each image just before the polyp information.

In addition, an error that caused a delay in downloading a snapshot is
corrected. For this, a function is created that assigns the image data
to the html download element only when it is clicked on.