Disables video upload if no files have been selected
Before, a required was used but it didn't work, therefore, it allowed to
upload a video without selecting a file. This made the video always stay
loading in the view. Now, the change event is used to change the
videoUploaded variable to true at the time the file has been selected.
This allows to avoid uploading a video without a selected file.