Fixes the percentage of the galleries
This commit fixes the way in which the percentage is calculated so
as not to exceed 100%. The problem originated because there were images
located without an associated polyp, and the percentages were obtained
based on located images about images with polyp. Now, images located
without associated polyp are also considered to calculate the
percentage.