Scroll to top when a video segment is played
Currently, when a video segment (polyp or modification) was played in
the video edition view, the video could be outside the viewport
(specially when the segment list is too long), causing the user to
miss a few seconds of the segment played.

This commit adds an automatic scroll to the top of the view, so that
the video player is displayed automatically when a segment is played.