Allows creating single-second video fragments
The creation of video fragment (polyp or modifier) was currently
restricted to fragments with a finishing second later to the starter
second. However, with current interpretation of starting and
finishing seconds, where the former represents the first frame of the
second while the latter represents the last frame of the second, it
makes sense to have a fragment with the same second for these two
instants.

Therefore, this commit removes the limitation of having different
starting and finishing seconds in the video fragments.