Adds navigation to the polyp recording view
The polyp recording view now allows navigating the polyp recordings
of a polyp dataset.

This commit includes several changes on the video component, as a
smoother progress bar advance of an improved interval size
calculation.

There is also a bug that makes the navigation randomly crash in
Firefox browsers that should be fixed in the future.