Fixes time calculation and polyp info presentation in polyp recording
The time calculations in the polyp recording view were not properly
done and failed in some limit cases. This commit fixes these errors
and now time to frame and frame to time conversions are uniformized.

In addition, the polyp information was shown in this view only when
the current frame contained an image, which was an error. Now, polyp
information is always shown as long as the current polyp recording
has an associated polyp.