Moves manually selected field in polyp recording view
In the polyp recording view, the manually selected image information
is now shown in the image table instead of as part of the polyp
information panel. This information was initially placed in this
panel because it was taken form the image view component, in which
made sense. However, in this view, it make more sense to show it as
part of the images table.