Allows sorting polyp recordings in dataset by images
The polyp recordings in a dataset can now be ordered by the number of
images in that belong to the gallery of the dataset. This is
important, as ordering by image is very useful for users annotating
polyp recordings.

When the polyp recordings are sorted, the polyp recording view
navigation uses the same order to navigate the polyp recordings.