Fixes and simplifies gallery image stats
The image stats of the galleries (i.e. total images, images with
location, etc.) was confusing and difficult to understand to the
users, as it differentiates between images with and without polyp,
although there should not exists images without polyp in the
database.

This commit fixes and simplifies these stats, mainly, by hiding to
the final user the existence of images without polyp.