Fixes bug in add polyp from video editor
The creation of polyps from the video editor view was failing due to
a change in the type of the exploration linked to the video under
edition, as it was changed in recent commits from a string to an
Exploration.

This commit fixes this bug by taking this type change into account
and, additionally, simplifies the query to the backend, as,
currently, the exploration object was retrieved before requesting
the creation of the polyp and now only the later request is needed.