Fixes image count in polyp dataset's recording list
The image count in the polyp recording list of the polyp dataset
view was not working correctly, as the images in the last second of
the recording was not taken into account unless they were exactly
at the start of the second.

This commit fixes this error and now the images are correctly
counted.